+++ express-zod-api/src/config-type.ts
@@ -156,12 +156,11 @@ interface GracefulOptions {
   beforeExit?: () => void | Promise<void>;
 }
 
-/** @todo consider removing Promise to make createServer sync in next major */
 type ServerHook = (params: {
   app: IRouter;
   /** @desc Returns child logger for the given request (if configured) or the configured logger otherwise */
   getLogger: GetLogger;
-}) => void | Promise<void>;
+}) => void;
 
 export interface HttpConfig {
   /** @desc Port, UNIX socket or custom options. */