+++ express-zod-api/src/documentation.ts
@@ -1,12 +1,14 @@
 import {
+  type InfoObject,
   type OperationObject,
   type ReferenceObject,
   type ResponsesObject,
-  type SchemaObject,
+  type SchemaObjectValue,
   type SecuritySchemeObject,
   type SecuritySchemeType,
+  type ServerObject,
   OpenApiBuilder,
-} from "openapi3-ts/oas31";
+} from "openapi3-ts/oas32";
 import * as R from "ramda";
 import { type ResponseVariant, responseVariants } from "./api-response";
 import { contentTypes } from "./content-type";
@@ -59,9 +61,14 @@ const defaultSummarizer: Summarizer = ({
 }) => trim(summary);
 
 interface DocumentationParams {
-  title: string;
-  version: string;
-  serverUrl: string | [string, ...string[]];
+  /** @desc At least title and version properties are required */
+  info: InfoObject;
+  /** @desc Server URL(s) or their complete definitions */
+  server:
+    | string
+    | [string, ...string[]]
+    | ServerObject
+    | [ServerObject, ...ServerObject[]];
   routing: Routing;
   config: CommonConfig;
   /**
@@ -113,7 +120,7 @@ export class Documentation extends OpenApiBuilder {
 
   #makeRef(
     key: object | string,
-    value: SchemaObject | ReferenceObject,
+    value: SchemaObjectValue | ReferenceObject,
     proposedName?: string,
   ): ReferenceObject {
     let name = this.#references.get(key); // search in the cache by the given key
@@ -166,11 +173,11 @@ export class Documentation extends OpenApiBuilder {
     return `${subject.type.toUpperCase()}_${nextId}`;
   }
 
-  #addMetadata({ title, version, serverUrl, tags }: DocumentationParams) {
-    this.addInfo({ title, version });
-    for (const url of typeof serverUrl === "string" ? [serverUrl] : serverUrl)
-      this.addServer({ url });
+  #addMetadata({ tags, info, server }: DocumentationParams) {
+    this.addInfo(info);
     if (tags) this.rootDoc.tags = depictTags(tags);
+    for (const one of Array.isArray(server) ? server : [server])
+      this.addServer(typeof one === "string" ? { url: one } : one);
   }
 
   #makeEndpointHandler({